requirements.txt
face_recognition==1.3.0
Flask==2.3.2
flask_openapi3==2.4.0
numpy==1.25.0
Pillow==10.0.0
pydantic==1.10.11
opencv-python-headless
gunicorn